Delay template filters until bp_before_group_plugin_template action.
A couple of filters were being added early enough in the group render routine that an extra "create" form was being added in response to BP setting the post ID for the group's page to 0. This changeset delays adding those filters until the `bp_before_group_plugin_template` action so that the doc filters do not take over the display of the main post on the page.